+++ src/containers/Blog/Blog.css
@@ -1,7 +1,23 @@
-.Posts {
-    display: flex;
-    flex-flow: row wrap;
-    justify-content: center;
-    width: 80%;
-    margin: auto;
+.Blog ul {
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    width: 100%;
+    text-align: center;
+}
+
+.Blog li {
+    margin: 20px;
+    display: inline-block;
+}
+
+.Blog a {
+    color: black;
+    text-decoration: none;
+}
+
+.Blog a:hover,
+.Blog a:active,
+.Blog a.active {
+    color: #fa923f;
 }
