+++ src/containers/Blog/Blog.js
@@ -1,60 +1,26 @@
 import React, { Component } from 'react';
-import axios from 'axios';
-import Post from '../../components/Post/Post';
-import FullPost from '../../components/FullPost/FullPost';
-import NewPost from '../../components/NewPost/NewPost';
+import Posts from '../../containers/Blog/Posts/Posts'; 
 import './Blog.css';
+import { Route, NavLink } from 'react-router-dom';
+import NewPost from './NewPost/NewPost';
 
-class Blog extends Component {
-    state = {
-        posts: [],
-        selectedPostId: null
-    }
-    componentDidMount() {
-        axios.get('/posts')
-        .then(response => {
-            const posts = response.data.slice(0,4);
-            const updatedPosts = posts.map(post => {
-                return {
-                    ...post,
-                    author: 'SK'
-                }
-            })
-            this.setState({
-                posts: updatedPosts
-            });
-            // console.log(response);
-        })
-        .catch((err) => {
-            console.log(err);
-        });
-    }
-
-    postClickHandler = (id) => {
-        this.setState({
-            selectedPostId: id
-        });
-    }
+class Blog extends Component { 
     render () {
-        const posts = this.state.posts.map((post) => {
-            return <Post 
-            title = {post.title} 
-            author = {post.author} 
-            key = {post.id}
-            clicked = {() => {this.postClickHandler(post.id)}}></Post>;
-        })
-
         return (
             <div>
-                <section className="Posts">
-                    {posts}
-                </section>
-                <section>
-                    <FullPost id = {this.state.selectedPostId} />
-                </section>
-                <section>
-                    <NewPost />
-                </section>
+                <header className = "Blog">
+                    <nav>
+                        <ul>
+                            <li><NavLink to = "/" exact >Home</NavLink></li>
+                            <li><NavLink to = {{
+                                pathname: '/new-post'
+                            }}>New Post</NavLink></li>
+                        </ul>
+                    </nav>
+                </header>
+                {/* <Route path = "/" exact render = {() => <h1>Home</h1>}></Route> */}
+                <Route path = "/" exact component = {Posts}></Route>
+                <Route path = "/new-post" exact component = {NewPost}></Route>
             </div>
         );
     }
