+++ src/containers/Blog/Posts/Posts.css
@@ -1,0 +1,7 @@
+.Posts {
+    display: flex;
+    flex-flow: row wrap;
+    justify-content: center;
+    width: 80%;
+    margin: auto;
+}
